requirements.txt
opencv-python==4.5.5.64
numpy<2.0.0
yt-dlp>=2023.3.4
torch>=1.10.0
torchvision>=0.11.0
flask>=2.0.0
playwright>=1.30.0
python-dotenv>=0.19.0
